requirements.txt
Django==1.3.1
PIL==1.1.7
South==0.7.3
Whoosh==2.3.0
distribute==0.6.19
django-auth-ldap==1.0.11
django-extensions==0.7.1
django-filebrowser==3.4.2
django-grappelli==2.3.8
django-haystack==1.2.7
python-ldap==2.3.13
wsgiref==0.1.2
gunicorn==0.13.4
django-context-extras==0.2.4
openpyxl==1.5.7
Celery==2.4.4
django-celery==2.4.2
geopy==0.94.1
requests==0.8.2
django-kombu==0.9.4
docutils==0.8.1
psycopg2==2.4.4
django-tinymce
django-flatpages-tinymce
-e git+git://github.com/alex/django-filter.git@51b39fca089f2cb1d3d3501d2e164635fbe52864#egg=django-filter
django-filter==0.5.3
django-tables2==0.10.0
easy-thumbnails==1.0-alpha-21
